* master: (48 commits)
  Threshold filter gets the right range.
  Added missing include.
  Mapper automatically uses the range from its associated lookup table.
  Fixed threshold filter color lookup.
  Fixed threshold filter lookup table coloring.
  Added -o switch for additional CMake options.
  fixed a bug in rotateMesh program (deleted an object twice)
  fixed a normalisation thingy
  changed distance computation in quadtree fixed a compile error in OGSError
  - Changed reading gli files, the function returns now a std::vector<std::string>. In this vector of strings possible error messages are catched and can be desplayed later on. - Changed test if point is already in quadtree.
  just some cosmetic changes...
  Reovery MarkInterface_mHM_Hydro_3D() for marking nodes on the top surface
  Fixed empty VtkAlgorithmProperties in point set item.
  added an "epsilon" to the search of mesh nodes within surfaces
  Correct the getElementFaces...
  some fixes for surface integration via GUI
  added an optional parameter to rotateMesh program
  Fixed css link on doxygen.
  Removed debug output.
  modified rotation programm
  ...

+++ VtkVis/VtkRaster.cpp
@@ -208,7 +208,7 @@ float* VtkRaster::loadDataFromASC(const std::string &fileName,
 
 		// shift noData values into normal pixel-range and set transparancy values for all pixels
 		size_t nPixels = header.ncols * header.nrows;
-		for (int j = 0; j <nPixels; j++)
+		for (size_t j = 0; j < nPixels; j++)
 		{
 			if (values[j*2] == header.noData)
 			{
